first_mrp

+++ database/migrations/2024_12_02_015543_create_open_pr_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('old_name');
             $table->date('pr_date');
             $table->date('request_date');
-            $table->integer('qty');
+            $table->float('qty');
             $table->string('pr_status');
             $table->date('date');
             $table->timestamps();
